refactor: Split Redis cache from underlying client
Redis can be used independently for non-cache services. This gives a
better structure and cleans up some tech debt.
