+++ scripts/simulateFill.ts
@@ -4,7 +4,6 @@ import { LogDescription } from "@ethersproject/abi";
 import { CHAIN_IDs } from "@across-protocol/constants";
 import { RelayData } from "../src/interfaces";
 import {
-  disconnectRedisClients,
   EvmAddress,
   getNetworkName,
   getProvider,
@@ -16,6 +15,7 @@ import {
   getBlockForTimestamp,
   winston,
 } from "../src/utils";
+import { disconnectRedisClients } from "../src/utils/redis";
 import * as utils from "./utils";
 
 const { NODE_SUCCESS, NODE_INPUT_ERR, NODE_APP_ERR } = utils;
