+++ scripts/swapOnBinance.ts
@@ -54,7 +54,6 @@ import {
   getNetworkName,
   getNativeTokenInfoForChain,
   getProvider,
-  getRedisCache,
   getSolanaTokenBalance,
   getSvmProvider,
   getTokenInfo,
@@ -79,6 +78,7 @@ import {
   willSucceed,
   readableBinanceWithdrawalStatus,
 } from "../src/utils";
+import { getRedisCache } from "../src/cache/redis";
 import { getCloidForAccount } from "../src/rebalancer/utils/utils";
 
 // Time between polling attempts when waiting on asynchronous operations to complete.
