+++ src/clients/TokenClient.ts
@@ -1,6 +1,7 @@
 import { utils as sdkUtils } from "@across-protocol/sdk";
 import { HubPoolClient, SpokePoolClient, SpokePoolManager, TransactionClient } from ".";
 import { CachingMechanismInterface, L1Token, Deposit } from "../interfaces";
+import { getRedisCache } from "../cache/redis";
 import {
   BigNumber,
   bnZero,
@@ -18,7 +19,6 @@ import {
   submitTransaction,
   toBN,
   winston,
-  getRedisCache,
   getInventoryBalanceContributorTokens,
   getRemoteTokenForL1Token,
   getTokenInfo,
