+++ src/common/ClientHelper.ts
@@ -13,7 +13,6 @@ import {
   getCurrentTime,
   SpokePool,
   isDefined,
-  getRedisCache,
   getArweaveJWKSigner,
   chainIsEvm,
   forEachAsync,
@@ -30,6 +29,7 @@ import {
   SpokePoolClient,
 } from "../clients";
 import { CommonConfig } from "./Config";
+import { getRedisCache } from "../cache/redis";
 import { SpokePoolClientsByChain } from "../interfaces";
 import { caching, clients, arch } from "@across-protocol/sdk";
 import V3_SPOKE_POOL_ABI from "./abi/V3SpokePool.json";
