+++ src/deposit-address/index.ts
@@ -1,4 +1,5 @@
-import { winston, config, startupLogLevel, Signer, disconnectRedisClients, getDispatcherKeys } from "../utils";
+import { winston, config, startupLogLevel, Signer, getDispatcherKeys } from "../utils";
+import { disconnectRedisClients } from "../utils/redis";
 import { DepositAddressHandler } from "./DepositAddressHandler";
 import { DepositAddressHandlerConfig } from "./DepositAddressHandlerConfig";
