+++ src/finalizer/index.ts
@@ -16,7 +16,6 @@ import {
   Signer,
   blockExplorerLink,
   config as dotenvConfig,
-  disconnectRedisClients,
   getMultisender,
   getNetworkName,
   Multicall2Call,
@@ -31,6 +30,7 @@ import {
   getProvider,
   chunk,
 } from "../utils";
+import { disconnectRedisClients } from "../utils/redis";
 import { ChainFinalizer, CrossChainMessage, Finalizer, isAugmentedTransaction } from "./types";
 import {
   arbStackFinalizer,
