+++ src/finalizer/utils/opStack.ts
@@ -21,7 +21,6 @@ import {
   getCachedProvider,
   getCurrentTime,
   getNetworkName,
-  getRedisCache,
   getUniqueLogIndex,
   getViemChain,
   groupObjectCountsByProp,
@@ -45,6 +44,7 @@ import {
   EvmAddress,
   ZERO_ADDRESS,
 } from "../../utils";
+import { getRedisCache } from "../../cache/redis";
 import {
   CONTRACT_ADDRESSES,
   OPSTACK_CONTRACT_OVERRIDES,
