+++ src/finalizer/utils/polygon.ts
@@ -10,7 +10,6 @@ import {
   getCachedProvider,
   getUniqueLogIndex,
   getCurrentTime,
-  getRedisCache,
   getBlockForTimestamp,
   Multicall2Call,
   TOKEN_SYMBOLS_MAP,
@@ -23,6 +22,7 @@ import {
   assert,
   isDefined,
 } from "../../utils";
+import { getRedisCache } from "../../cache/redis";
 import { EthersError, TokensBridged } from "../../interfaces";
 import { HubPoolClient, SpokePoolClient } from "../../clients";
 import { FinalizerPromise, CrossChainMessage } from "../types";
