+++ src/finalizer/utils/zkSync.ts
@@ -7,7 +7,6 @@ import {
   convertFromWei,
   getBlockForTimestamp,
   getCurrentTime,
-  getRedisCache,
   getTokenInfo,
   getUniqueLogIndex,
   Multicall2Call,
@@ -23,6 +22,7 @@ import {
   Provider,
   CHAIN_IDs,
 } from "../../utils";
+import { getRedisCache } from "../../cache/redis";
 import { FinalizerPromise, CrossChainMessage } from "../types";
 
 type TokensBridged = interfaces.TokensBridged;
