+++ src/libexec/RelayerSpokePoolListener.ts
@@ -4,7 +4,6 @@ import { Contract, utils as ethersUtils } from "ethers";
 import * as utils from "../../scripts/utils";
 import { EventListener } from "../clients";
 import {
-  disconnectRedisClients,
   exit,
   isDefined,
   getBlockForTimestamp,
@@ -13,11 +12,12 @@ import {
   getNetworkName,
   getProvider,
   getSpokePool,
-  getRedisCache,
   Logger,
   Provider,
   winston,
 } from "../utils";
+import { getRedisCache } from "../cache/redis";
+import { disconnectRedisClients } from "../utils/redis";
 import { ScraperOpts } from "./types";
 import { postBlock, postEvents, removeEvent } from "./util/ipc";
 import { scrapeEvents as _scrapeEvents } from "./util/evm";
