+++ src/libexec/RelayerSpokePoolListenerTVM.ts
@@ -5,7 +5,6 @@ import { AbiEvent, BaseError, Block, createPublicClient, http, Log as viemLog, p
 import * as utils from "../../scripts/utils";
 import { Log } from "../interfaces";
 import {
-  disconnectRedisClients,
   EventManager,
   exit,
   isDefined,
@@ -18,13 +17,14 @@ import {
   getOriginFromURL,
   getProvider,
   getProviderHeaders,
-  getRedisCache,
   getViemChain,
   Logger,
   Provider,
   SpokePool,
   winston,
 } from "../utils";
+import { getRedisCache } from "../cache/redis";
+import { disconnectRedisClients } from "../utils/redis";
 import { ScraperOpts } from "./types";
 import { postBlock, postEvents, removeEvent } from "./util/ipc";
 import { scrapeEvents as _scrapeEvents } from "./util/evm";
