+++ src/monitor/index.ts
@@ -1,4 +1,5 @@
-import { winston, processEndPollingLoop, config, startupLogLevel, Signer, disconnectRedisClients } from "../utils";
+import { winston, processEndPollingLoop, config, startupLogLevel, Signer } from "../utils";
+import { disconnectRedisClients } from "../utils/redis";
 import { Monitor } from "./Monitor";
 import { MonitorConfig } from "./MonitorConfig";
 import { constructMonitorClients } from "./MonitorClientHelper";
