+++ src/refiller/Refiller.ts
@@ -16,7 +16,6 @@ import {
   getNativeTokenAddressForChain,
   getNativeTokenSymbol,
   getNetworkName,
-  getRedisCache,
   getSolanaTokenBalance,
   getSvmProvider,
   getTokenInfo,
@@ -44,7 +43,7 @@ import { SWAP_ROUTES, SwapRoute, CUSTOM_BRIDGE, CANONICAL_BRIDGE } from "../comm
 import ERC20_ABI from "../common/abi/MinimalERC20.json";
 import { arch } from "@across-protocol/sdk";
 import { AcrossSwapApiClient, BalanceAllocator, MultiCallerClient, TransactionClient } from "../clients";
-import { RedisCache } from "../caching/RedisCache";
+import { getRedisCache, RedisCache } from "../cache/redis";
 
 export interface RefillerClients {
   balanceAllocator: BalanceAllocator;
