+++ src/scripts/validateRunningBalances.ts
@@ -45,7 +45,6 @@ import {
   sortEventsDescending,
   paginatedEventQuery,
   ZERO_ADDRESS,
-  disconnectRedisClients,
   Signer,
   getSigner,
   getEndBlockBuffers,
@@ -56,6 +55,7 @@ import {
   isDefined,
   BigNumber,
 } from "../utils";
+import { disconnectRedisClients } from "../utils/redis";
 import { createDataworker } from "../dataworker";
 import { _buildSlowRelayRoot, getBlockForChain } from "../dataworker/DataworkerUtils";
 import { Log, ProposedRootBundle, SpokePoolClientsByChain, BundleData } from "../interfaces";
